fix yet another froxelization bug
This came up when trying to use a very small number of froxels like
1x1x1 (just one!). 

Some froxels where not detected as used when scanning horizontally,
this would happen when a sphere was fully inside a froxel (i.e. not
actually intersecting any planes), and this was the very last
column on the right.

Now we use the same code we use for Z and Y (which where working fine)
for the X loop.